Rename: 폴더 정리
panel 폴더 제거
assets 절대경로 추가
page로 되어있던 절대경로 pages로 변경
