--- src/storybook/page/main/Title.stories.jsx
@@ -1,12 +1,0 @@
-import Title from '@/page/main/panel/Title';
-
-const meta = {
-    title: 'Main/Title',
-    component: Title,
-};
-
-export default meta;
-
-const Default = {};
-
-export { Default };
